+++ weaver/wps_restapi/swagger_definitions.py
@@ -3978,6 +3978,18 @@ class NotFoundJobResponseSchema(ExtendedMappingSchema):
     body = ErrorJsonResponseBodySchema()
 
 
+class GoneJobResponseSchema(ExtendedMappingSchema):
+    description = "Job reference UUID cannot be dismissed again or its result artifacts were removed."
+    examples = {
+        "JobDismissed": {
+            "summary": "Example response when specified job reference was already dismissed.",
+            "value": EXAMPLES["job_dismissed_error.json"]
+        }
+    }
+    header = ResponseHeaders()
+    body = ErrorJsonResponseBodySchema()
+
+
 class OkGetJobInputsResponse(ExtendedMappingSchema):
     header = ResponseHeaders()
     body = JobInputsSchema()
@@ -4226,6 +4238,7 @@ get_prov_single_job_status_responses.update({
 delete_job_responses = {
     "200": OkDismissJobResponse(description="success"),
     "404": NotFoundJobResponseSchema(),
+    "410": GoneJobResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
 delete_prov_job_responses = copy(delete_job_responses)
@@ -4254,6 +4267,7 @@ get_job_outputs_responses = {
         }
     }),
     "404": NotFoundJobResponseSchema(),
+    "410": GoneJobResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
 get_prov_outputs_responses = copy(get_job_outputs_responses)
@@ -4271,6 +4285,7 @@ get_job_results_responses = {
         }
     }),
     "404": NotFoundJobResponseSchema(),
+    "410": GoneJobResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
 get_prov_results_responses = copy(get_job_results_responses)
@@ -4285,6 +4300,7 @@ get_exceptions_responses = {
         }
     }),
     "404": NotFoundJobResponseSchema(),
+    "410": GoneJobResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
 get_prov_exceptions_responses = copy(get_exceptions_responses)
@@ -4299,6 +4315,7 @@ get_logs_responses = {
         }
     }),
     "404": NotFoundJobResponseSchema(),
+    "410": GoneJobResponseSchema(),
     "500": InternalServerErrorResponseSchema(),
 }
 get_prov_logs_responses = copy(get_logs_responses)